requirements.txt
ase==3.22.1
nequip==0.5.6
digital-experiments==1.1.4
notebook==6.5.2
load-atoms==0.0.12
quippy-ase==0.9.12
git+https://github.com/ICAMS/python-ace.git
locache==3.0.0
rich==13.3.5
ipywidgets==8.0.6
umap-learn==0.5.3